refactor: replace ThemeContext.Provider with ThemeContextProvider component
