+++ src/components/Providers/index.tsx
@@ -1,38 +1,19 @@
 "use client"
 
-import {ReactNode, useEffect, useState} from "react"
-import THEMES from "@/THEMES.json"
-import ThemeContext from "@/context/ThemeContext"
-import {TThemeName} from "@/types/TThemeName"
+import {ReactNode} from "react"
+import ThemeContextProvider from "./ThemeContextProvider"
 
 type TProvidersProps = {
     children: ReactNode
 }
 
 function Providers({children}: TProvidersProps) {
-    const [currentThemeName, setCurrentThemeName] = useState<TThemeName>("default")
-    const [currentTheme, setCurrentTheme] = useState(THEMES[currentThemeName])
 
-    function changeThemeTo(newThemeName: TThemeName): void {
-        setCurrentThemeName(newThemeName)
-        setCurrentTheme(THEMES[newThemeName])
-    }
-
-    useEffect(() => {
-        changeThemeTo(currentThemeName)
-    }, [currentThemeName])
-
-    const appState = {
-        currentTheme,
-        changeThemeTo: changeThemeTo,
-    }
 
     return (
-        <ThemeContext.Provider value={appState}>
+        <ThemeContextProvider>
             {children}
-        </ThemeContext.Provider>
+        </ThemeContextProvider>
     )
 }
 export default Providers
-
-export {ThemeContext}
